[SSHD-816] Moved all 'main' code for client commands to sshd-cli module
