+++ pom.xml
@@ -1010,6 +1010,7 @@
         <module>sshd-git</module>
         <module>sshd-contrib</module>
         <module>sshd-spring-sftp</module>
+        <module>sshd-cli</module>
         <module>assembly</module>
     </modules>
 </project>
