+++ sshd-cli/src/test/java/org/apache/sshd/cli/client/ScpCommandMainDevelopment.java
@@ -17,20 +17,20 @@
  * under the License.
  */
 
-package org.apache.sshd.client.scp;
+package org.apache.sshd.cli.client;
 
 /**
- * Just a test class used to invoke {@link org.apache.sshd.client.scp.DefaultScpClient#main(String[])} in
+ * Just a test class used to invoke {@link ScpCommandMain#main(String[])} in
  * order to have logging - which is in {@code test} scope
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public final class ScpCommandMain {
-    private ScpCommandMain() {
+public final class ScpCommandMainDevelopment {
+    private ScpCommandMainDevelopment() {
         throw new UnsupportedOperationException("No instance");
     }
 
     public static void main(String[] args) throws Exception {
-        DefaultScpClient.main(args);
+        ScpCommandMain.main(args);
     }
 }
