+++ sshd-cli/src/test/java/org/apache/sshd/cli/client/SftpCommandMainDevelopment.java
@@ -17,20 +17,20 @@
  * under the License.
  */
 
-package org.apache.sshd.client.subsystem.sftp;
+package org.apache.sshd.cli.client;
 
 /**
- * Just a test class used to invoke {@link SftpCommand#main(String[])} in
+ * Just a test class used to invoke {@link SftpCommandMain#main(String[])} in
  * order to have logging - which is in {@code test} scope
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public final class SftpCommandMain {
-    private SftpCommandMain() {
+public final class SftpCommandMainDevelopment {
+    private SftpCommandMainDevelopment() {
         throw new UnsupportedOperationException("No instance");
     }
 
     public static void main(String[] args) throws Exception {
-        SftpCommand.main(args);
+        SftpCommandMain.main(args);
     }
 }
